editor 사용 전면 개선/게시물 삭제,수정 권한 부여 기능 개선/로그인 후 페이지 이동 개선 등

+++ WebContent/jsp/article/detail.jsp
@@ -1,4 +1,5 @@
 <%@ include file="/jsp/part/head.jspf"%>
+<%@ include file="/jsp/part/toastUIEditor.jspf"%>
 <%@ page import="com.sbs.java.blog.dto.Article"%>
 <%@ page import="com.sbs.java.blog.dto.ArticleReply"%>
 <%@ page import="java.util.List"%>
@@ -6,47 +7,6 @@
 	pageEncoding="UTF-8"%>
 
 
-<!-- 마크다운 뷰어에 필요한 코드.. (참고: 딱 필요한 것만 추리지 못함) -->
-
-<!-- 하이라이트 라이브러리 추가, 토스트 UI 에디터에서 사용됨 -->
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/highlight.min.js"></script>
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/styles/default.min.css">
-
-<!-- 하이라이트 라이브러리, 언어 -->
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/css.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/javascript.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/java.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/xml.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/php.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/php-template.min.js"></script>
-<script
-	src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/languages/sql.min.js"></script>
-
-<!-- 코드 미러 라이브러리 추가, 토스트 UI 에디터에서 사용됨 -->
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.48.4/codemirror.min.css" />
-
-<!-- 토스트 UI 에디터, 자바스크립트 코어 -->
-<script
-	src="https://uicdn.toast.com/editor/latest/toastui-editor-viewer.min.js"></script>
-
-<!-- 토스트 UI 에디터, 신택스 하이라이트 플러그인 추가 -->
-<script
-	src="https://uicdn.toast.com/editor-plugin-code-syntax-highlight/latest/toastui-editor-plugin-code-syntax-highlight-all.min.js"></script>
-
-<!-- 토스트 UI 에디터, CSS 코어 -->
-<link rel="stylesheet"
-	href="https://uicdn.toast.com/editor/latest/toastui-editor.min.css" />
-
-<!--  마크다운 뷰어 코드 끝 -->
 
 
 
@@ -88,25 +48,17 @@
 			<!-- 카테고리 게시물 접속했을 때, 목록 클릭하면 최신 게시물을 불러왔음. 카테고리 게시물로 이동하게 수정한 코드 -->
 			<a href="list?cateItemId=${param.cateItemId}&page=${param.page}"
 				class="back-icon list-icon"><i class="fas fa-arrow-left">목록
-			</i></a> <a
-				href="${pageContext.request.contextPath}/s/article/modify?id=<%=article.getId()%>"
+			</i></a>
+			<% if ( (boolean)article.getExtra().get("modifyAvailable"))  { %>
+				<a 	href="${pageContext.request.contextPath}/s/article/modify?id=<%=article.getId()%>"
 				class="back-icon  modify-icon"><i class="fas fa-edit">수정 </i></a>
+			<%} %>
+			
 
 		</div>
 		<div class="editor-box">
-			<script type="text/x-template" id="origin1" style="display: none;"><%=article.getBodyForXTemplate()%></script>
-			<div id="viewer1"></div>
-			<script>
-				var editor1__initialValue = getBodyFromXTemplate('#origin1') // trim() 추가했음. 
-				var editor1 = new toastui.Editor({
-					el : document.querySelector("#viewer1"),
-					viewer : true,
-					initialValue : editor1__initialValue, 
-// 					.replace(/<!--REPLACE:script-->/gi, 'script'),   common.js 로 옮겼음. (공용으로 사용하는)
-					plugins : [ toastui.Editor.plugin.codeSyntaxHighlight,
-							youtubePlugin, replPlugin, codepenPlugin ]
-				});
-			</script>
+			<script type="text/x-template" ><%=article.getBodyForXTemplate()%></script>
+			<div class="toast-editor toast-editor-viewer"></div>
 			<br />
 			<div class="move-button">
 				<%
@@ -132,9 +84,12 @@
 
 			</div>
 			<div class="delete-button">
-				<button type="submit"
-					onclick="location.href='doDelete?id=${param.id}'"
-					style="position: absolute; left:55%; top: 30%;">게시물 삭제</button>
+				<% if ( (boolean)article.getExtra().get("deleteAvailable"))  { %>
+					<button type="submit"  onclick="if (confirm('삭제하시겠습니까?') == false ) return false; location.href='doDelete?id=${param.id}'" 	style="position: absolute; left:55%; top: 30%;">
+					게시물 삭제 
+					</button>
+				<%} %>
+				
 			</div>
 		</div>
